Moved speaker test code to it's own source file
